# openapi.yaml
openapi: 3.0.0
info:
  title: 'Q Laravel Skeleton API'
  version: '0.9'
paths:
...
  /api/v1/publishers:
    get:
      tags:
        - Publishers
      summary: 'Get list of publishers'
      operationId: getPublishers
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                properties:
                  publisher: { $ref: '#/components/schemas/PublisherResource' }
                type: object
        '400':
          description: 'Bad Request'
        '401':
          description: Unauthorized
      security:
        -
          sanctum: []
    post:
      tags:
        - Publishers
      summary: 'Create new publisher'
      operationId: createPublisher
      requestBody:
        description: 'Create new publisher'
        required: true
        content:
          application/json:
            schema:
              required:
                - name
              properties:
                name:
                  type: string
                  example: 'Q Publishing'
                address:
                  type: string
                  example: 'Some address 33'
              type: object
      responses:
        '201':
          description: 'Publisher created successfully'
          content:
            application/json:
              schema:
                properties:
                  publisher: { $ref: '#/components/schemas/PublisherResource' }
                type: object
        '400':
          description: 'Bad Request'
        '401':
          description: Unauthorized
        '422':
          description: 'Validation failed'
      security:
        -
          sanctum: []
...
  '/api/v1/users/{user}/tags/{tag}':
    post:
      tags:
        - Users
      summary: 'Add tag for user'
      operationId: addUserTag
      parameters:
        -
          name: user
          in: path
          description: 'User id'
          required: true
          example: 1
        -
          name: tag
          in: path
          description: 'Tag id'
          required: true
          example: 1
      responses:
        '200':
          description: 'Tag added successfully'
          content:
            application/json:
              schema:
                properties:
                  user: { $ref: '#/components/schemas/UserResource' }
                type: object
        '400':
          description: 'Bad Request'
        '404':
          description: 'Not found'
      security:
        -
          sanctum: []
    delete:
      tags:
        - Users
      summary: 'Remove tag from user'
      operationId: removeUser
      parameters:
        -
          name: user
          in: path
          description: 'User id'
          required: true
          example: 1
        -
          name: tag
          in: path
          description: 'Tag id'
          required: true
          example: 1
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                properties:
                  user: { $ref: '#/components/schemas/UserResource' }
                type: object
        '400':
          description: 'Bad Request'
        '401':
          description: Unauthorized
        '404':
          description: 'Not found'
      security:
        -
          sanctum: []
  /api/v1/users:
    get:
      tags:
        - Users
      summary: 'Get list of users'
      operationId: getUsersV1
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                properties:
                  user: { $ref: '#/components/schemas/UserResource' }
                type: object
        '400':
          description: 'Bad Request'
        '401':
          description: Unauthorized
      security:
        -
          sanctum: []
    post:
      tags:
        - Users
      summary: 'Create new user'
      operationId: createUserV1
      requestBody:
        description: 'Create new user'
        required: true
        content:
          application/json:
            schema:
              required:
                - first_name
                - last_name
                - email
                - password
              properties:
                first_name:
                  type: string
                  example: John
                last_name:
                  type: string
                  example: Doe
                email:
                  type: string
                  example: john.doe@q.agency
                password:
                  type: string
                  example: secretPassword
              type: object
      responses:
        '201':
          description: 'User created successfully'
          content:
            application/json:
              schema:
                properties:
                  user: { $ref: '#/components/schemas/UserResource' }
                type: object
        '400':
          description: 'Bad Request'
        '422':
          description: 'Validation failed'
  '/api/v1/users/{user}':
    get:
      tags:
        - Users
      summary: 'Get user data'
      operationId: getUserV1
      parameters:
        -
          name: user
          in: path
          description: 'User id'
          required: true
          example: 1
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                properties:
                  user: { $ref: '#/components/schemas/UserResource' }
                type: object
        '400':
          description: 'Bad Request'
        '401':
          description: Unauthorized
        '404':
          description: 'Not found'
      security:
        -
          sanctum: []
    put:
      tags:
        - Users
      summary: 'Update existing user'
      operationId: updateUserV1
      parameters:
        -
          name: user
          in: path
          description: 'User id'
          required: true
          example: 1
      requestBody:
        description: 'Update existing user'
        required: true
        content:
          application/json:
            schema:
              required:
                - first_name
                - last_name
                - email
                - password
              properties:
                first_name:
                  type: string
                  example: John
                last_name:
                  type: string
                  example: Doe
                email:
                  type: string
                  example: john.doe@q.agency
                password:
                  type: string
                  example: secretPassword
              type: object
      responses:
        '200':
          description: 'User updated successfully'
          content:
            application/json:
              schema:
                properties:
                  user: { $ref: '#/components/schemas/UserResource' }
                type: object
        '400':
          description: 'Bad Request'
        '401':
          description: Unauthorized
        '422':
          description: 'Validation failed'
        '404':
          description: 'Not found'
      security:
        -
          sanctum: []
    delete:
      tags:
        - Users
      summary: 'Delete user'
      operationId: deleteUserV1
      parameters:
        -
          name: user
          in: path
          description: 'User id'
          required: true
          example: 1
      responses:
        '204':
          description: Deleted
          content:
            application/json:
              schema:
                properties:
                  message: { type: string, example: 'User deleted.' }
                type: object
        '400':
          description: 'Bad Request'
        '401':
          description: Unauthorized
        '404':
          description: 'Not found'
      security:
        -
          sanctum: []
  /api/v2/users:
    get:
      tags:
        - Users
      summary: 'Get list of users'
      operationId: getUsersV2
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                properties:
                  user: { $ref: '#/components/schemas/UserResource' }
                type: object
        '400':
          description: 'Bad Request'
        '401':
          description: Unauthorized
      security:
        -
          sanctum: []
    post:
      tags:
        - Users
      summary: 'Create new user'
      operationId: createUserV2
      requestBody:
        description: 'Create new user'
        required: true
        content:
          application/json:
            schema:
              required:
                - first_name
                - last_name
                - email
                - password
              properties:
                first_name:
                  type: string
                  example: John
                last_name:
                  type: string
                  example: Doe
                email:
                  type: string
                  example: john.doe@q.agency
                password:
                  type: string
                  example: secretPassword
              type: object
      responses:
        '201':
          description: 'User created successfully'
          content:
            application/json:
              schema:
                properties:
                  user: { $ref: '#/components/schemas/UserResource' }
                type: object
        '400':
          description: 'Bad Request'
        '422':
          description: 'Validation failed'
  '/api/v2/users/{user}':
    get:
      tags:
        - Users
      summary: 'Get user data'
      operationId: getUserV2
      parameters:
        -
          name: user
          in: path
          description: 'User id'
          required: true
          example: 1
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                properties:
                  user: { $ref: '#/components/schemas/UserResource' }
                type: object
        '400':
          description: 'Bad Request'
        '401':
          description: Unauthorized
        '404':
          description: 'Not found'
      security:
        -
          sanctum: []
    put:
      tags:
        - Users
      summary: 'Update existing user'
      operationId: updateUserV2
      parameters:
        -
          name: user
          in: path
          description: 'User id'
          required: true
          example: 1
      requestBody:
        description: 'Update existing user'
        required: true
        content:
          application/json:
            schema:
              required:
                - first_name
                - last_name
                - email
                - password
              properties:
                first_name:
                  type: string
                  example: John
                last_name:
                  type: string
                  example: Doe
                email:
                  type: string
                  example: john.doe@q.agency
                password:
                  type: string
                  example: secretPassword
              type: object
      responses:
        '200':
          description: 'User updated successfully'
          content:
            application/json:
              schema:
                properties:
                  user: { $ref: '#/components/schemas/UserResource' }
                type: object
        '400':
          description: 'Bad Request'
        '401':
          description: Unauthorized
        '422':
          description: 'Validation failed'
        '404':
          description: 'Not found'
      security:
        -
          sanctum: []
    delete:
      tags:
        - Users
      summary: 'Delete user'
      operationId: deleteUserV2
      parameters:
        -
          name: user
          in: path
          description: 'User id'
          required: true
          example: 1
      responses:
        '204':
          description: Deleted
          content:
            application/json:
              schema:
                properties:
                  message: { type: string, example: 'User deleted.' }
                type: object
        '400':
          description: 'Bad Request'
        '401':
          description: Unauthorized
        '404':
          description: 'Not found'
      security:
        -
          sanctum: []
...
components:
  schemas:
...
    BookResource:
      description: 'Class BookResource'
      properties:
        id:
          type: integer
          default: 1
        isbn:
          type: string
          default: 978-92-95055-02-5
        title:
          type: string
          default: 'How to tie your shoes'
        release_date:
          type: string
          default: '1999-25-11'
        format:
          type: string
          default: PDF
        pages:
          type: integer
          default: '218'
        publisher:
          $ref: '#/components/schemas/PublisherResource'
      type: object
    PublisherResource:
      description: 'Class PublisherResource'
      properties:
        id:
          type: integer
          default: 1
        name:
          type: string
          default: 'Q Books Ltd'
        address:
          type: string
          default: '5th Avenue, LA'
        books:
          $ref: '#/components/schemas/BookResource'
      type: object
...
    UserResource:
      description: 'Class UserResource'
      properties:
        id:
          type: integer
          default: 1
        first_name:
          type: string
          default: John
        last_name:
          type: string
          default: Doe
        email:
          type: string
          default: john.doe@q.agency
      type: object
...